refactor: extract setup manifest checks
